Refactor server to separate API layer from server logic (#6)
Implements modular architecture separating concerns:

- Created api/alpacaService.js with 4 async functions (fetchBars, fetchQuote, fetchOptionsData, fetchOptionContract)
- Created config/alpaca.js for centralized Alpaca client initialization
- Created routes/ directory with separate files for stocks, options, and health endpoints
- Created websocket/streamManager.js to handle WebSocket connections and Alpaca streaming
- Refactored server/index.js from 437 lines to 46 lines (90% reduction)

Benefits:
- Single responsibility per module
- Improved testability through clear abstractions
- Easier to mock and test individual components
- Pluggable data source architecture
- Better error handling and debugging

All existing functionality preserved and tested.

Closes #6
